Enum.split implementation
Could not figure out how to implement with tail call optimization.  This
means the function is not very optimal and shouldn’t be used in
practice as it will overflow the stack for large enumerations.

Method of dealing with negative counts feels naive but it’s the easiest
to understand so I’m happy with it.